--- app/Models/Member.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class Member extends Model
-{
-    use HasFactory;
-	protected $table = 'members';
-	protected $primaryKey = 'id';
-    protected $fillable = [
-        'uid', 'nickname', 'email', 'password'
-    ];
-}
